+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangelog.com/en/1.0.0/) from version 0.9 onwards.
 
 ## 0.16 (unreleased)
+- **Implemented trend reversal candles**
+- **Fixed NaN in DXIndicator, MinusDIIndicator, PlusDIIndicator if there is no trend**
 
 ### Breaking
 - **Upgraded to Java 11**